fix: Fix request #31084 Tuleap must be able to import huge XML files
If the XML file to import is too huge (our test was done with a 200MB file),
then its content is truncated and errors are raised because the truncated content
is not a valid XML content.

Note: composer must be run because the namespace Tuleap\XML does not
match the folder names.

Change-Id: I7a60328a1b5dcfde71752fda4ef66a7ec6cc5e3f