fix(mw-standalone): Clean-up `initializing` instances
Under certain conditions, it was possible to activate a Standalone
MediaWiki despite data from legacy MediaWiki being there.

This commit will clean-up the said instances to be able to get back to
something that can be further worked on.

This forgeupgrade only purge the tuleap_instances table and project &
site admin should act for the next steps:
* Either the legacy mediawiki was not used, hence the sys admin should
  use `clean-unused-mw` script to purge the existing tables so the
  project admin can move on
* Either it was actually used, hence the Standalone MW service should be
  disabled.

Closes request #31062 System service can be wrongly used

Change-Id: Ibee2c11b82a07cc79ffcde5b772b071264f59236