feat: comprehensive history for invitation
Add the id of the invitation in project history events so that project
admins can correlate which invitations led to member added.

This comes with a new status for invitation to differentiate the
invitation that are completed from the one the user actually used to
register.

This change is reflected in both instrumentation and project history:
every completed invitation are logged. We do not log the used invitation
in project history, it does not seem to be really relevant for the site
admin to know which specific invitation led to user registration.

In a next contribution the status `completed` will be enforced in the
database to not keep invitations as `sent` forever.

Part of story #30021: Invite buddies to my Tuleap project

Change-Id: Ic9a00de132d0ecccc91ba2d5f7f054c6da57ade8