Add the various list fields to the "flat" representations of the report's artifacts
Added a fallback locale when `$GLOBALS['Language']` is not defined when
initializing a \PFUser. It was an annoyance when building a
MinimalUserRepresentation because an anonymous user with no locale is
created and `GlobalLanguageMock` does not cover data providers.

Part of story #31087: have flat json representation on `/tracker_reports/:id/artifacts`

Change-Id: I39b5667c427fd8beefeb2219b776f8f64e633afc