feat: Close request #31094 Jira import tool must deal with missing agile endpoint
Given a Jira platform without the agile endpoint activated,
then our Jira import tool must not fail and must import all the issueType
without any agile content.

Technical note: We can rely on a 404 error because if there is no board
matching the query we will have an empty JSON as a return and not a 404.

Change-Id: Ic2f6ee609e4293bad662b7591abd3e68a204c176