feat: duplicate date reminders
Create a new tracker based on an existing one and/or create a new
project from an existing (not XML) one, the date reminder configuration
are now duplicated.

Due to the data organisation for ugroups (a column with comma separated
values) the code is more complex than usual duplicate as it's not
possible to do the transformation at SQL level.

Closes request #31098 Project created from DB templates doesn't keep date reminder configuration

Change-Id: Ib828281306129fd403f8d5f71fd406bec9341967