Ease deployment of keycloak in dev environment
We can ease the process of setup and configuration of Keycloak in a
development environment by adding the keycloak entry directly in our
`docker-compose.yml` and importing a ready to use Keycloak realm.

The imported realm:
* Provides a client `tuleap-dev` with secret
  `TULEAP_DO_NOT_USE_THIS_IN_PRODUCTION`
* Provides a user `jdoe` with password `welcome0`

That way the developer does not have to open the keycloak admin console
to search for the client secret.

Please see updated doc/back-end/keycloak.md to configure your Tuleap dev
instance to use Keycloak.

=> You can now connect with the Keycloak provider by using the user
`jdoe`.

Note: Updating temporarily the `docker-compose.yml` as it is currently
explained in the documentation was not necessarily a good idea since it
leaves orphan keycloak container and docker-compose complains about it.

Part of request #31107: Ease deployment of keycloak in dev environment

Change-Id: Ic73c317bea485af3d75957fa06f5fb9a489cb9aa