fix: force reconnection to db during artifact import
If the XML file to import is huge (`389MB` archive including `241M` of
xml and file attachments), then the import stops itself during the
process. It looks like forcing the reconnection at each imported
artifact allows to complete the import.

Request #31084 already allows to import huge XML, but it is not enough
with attachments.

Closes request #31112: Tuleap must be able to import huge XML files

Change-Id: I3d7786b765f415d634361018f72f7972cbdd26da