chore: import developer documentation
In previous contribution, all Tuleap related code was imported in the
main source tree. In no longer make sense to maintain the documentation
outside the source tree.

This is the result of a mostly raw import of existing documentation with
pango to make the rst/md transcription. It required a lot of manual
fixing after conversion. It's likely not finished.

The relevancy or the organization of the documentation has not been
challenged during the import.

Part of: request #31116 Import Tuleap developers documentation in sources

Change-Id: I15237ce1acb0f911e9247b415cd1ff5d52b66c80