Have REST api types in a shared lib
part of story #31091 Add the hybrid comment to manage comments

Given the fact that we now have 2 apps and 1 component working with the
REST api types of the pullrequest plugin routes, it would be great to
stop duplicating the types definitions everywhere. Moreover, several
types used in both app are defined in lib/pullrequest-comments. It is not
really the place to do it, so let's move those definitions in dedicated
libraries.

-- No changes expected --

Change-Id: I300d97e92efe3555d8e252182ad643b74022a08a