pull-request comment skeleton when timeline is loading
part of story #31091 Add the hybrid comment to manage comments

This contribution introduces a skeleton version of the
pullrequest-comment component. I didnt use the tlp-skeleton color
because this variable is a RGBa color (RGB + alpha), and, since there
are overlapping elements in the component (ex: avatar on followup's
border), the opacities sum up, and thus makes the visual rendering
terrible.

How to test:
- When the timeline is loading, the skeleton is displayed.

Note:
- You can comment the line 140 in OverviewThreads to see it better.

Change-Id: I692fb0318758046b1d1a75ba5b6425ae03aa738b