Fix request #30398 Follow-up comments in modal sort is incorrect
Users can choose the order of the follow-up comments in Artifact View.
The order should be the same when follow-up comments are displayed in
the Artifact Modal.

The issue is here when the user does not have a user preference set for
the comment order in the tracker. In that case, before this patch, the
order would be inverted in the modal. It should now be the same as the
artifact view.

Change-Id: I2835290a22f863f8604f3204add98fbf94b84e38