Submit new comments
part of story #31091 Add the hybrid comment to manage comments

Note to reviewer:

I apologize for the huge size of this contribution. I had to see if it
was really possible to have a single component to handle the creation of
file-diff comments and global comments. Turned out that, yes, it was!

The good stuff is that when we'll work on the markdown feature, we will
not have to do crazy stuff in the angular app. Moreover, we would have
to homogenize the design between the two apps for this feature one day
or another, and I truely believe that it is better not to have
duplicated implementations here and there.

How to test:
- Pick up a pull-request with no comments
--> The empty state is displayed, it is nicely proportioned
--> The new comment component is displayed at the bottom of the section

- Write a comment and submit
--> The comment is added to the section
--> The content of the component is cleared
--> The empty state is no more displayed
--> The component is displayed right after the newly created comment
    (as per UXD's instructions)

- Throw a RestException in PullRequestsResource::1143
- Submit a new comment
--> The error modal opens

- Go to the [Changes] tab
--> No functional changes expected, you should still be able to submit
    inline comments as before.

Change-Id: Idc73c3cea5b8cf3287b975f29baea747ff92f0a2