refactor: UI should not know about link type representations
No functional change expected in the artifact link field. It still shows
the list of available link types with only "untyped", "parent" and
"child" types.

The UI adapter should not know about REST representations. It is the job
of the REST adapter. Thus, parsing the REST representation of link types
is moved to the REST adapter. It builds domain objects, that the UI
adapter displays as HTML.

part of story #24970 [modal] select type for links

Change-Id: I7cf2d9af8afe45f9bc3b2585c5c1ef053ffa5a44