Stop passing a RelativeDateHelper to the comment component
part of story #31091 Add the hybrid comment to manage comments

At the very beginning, the idea was to fix the case for the attribute
relativeDateHelper of the <tuleap-pullrequest-comment/> component, but
it turned out that it was breaking its usage in the angular app.

When I realized that the angular app used the same object (with type
definition comming from @tuleap/plugin-pullrequest-comments) to display
both its relative dates, and the component's relative dates, I've
decided to stop requiring a relative date helper in the component, and
build it in its controller instead.

As a reminder, we had initially implemented a RelativeDateHelper in
order to ease the initialisation of <tlp-relative-date/> inside the
angular app's templates.

Now:
- The overview app has its own PullRequestRelativeDate vue component.
- The angular app has its own relative dates helper.
- The comment component has its own too.

-- No functional change expected --

Change-Id: I9a90a07d3d6d803d20ad80043632f6b83970d773