fix request #31132: ignore already exported artifacts
Sometimes the generated XML content during an import from Jira contains
twice the same artifact, leading to errors during import.

We suspect that when doing an import from Jira that is taking many
seconds to complete, and there are some activities on the Jira instance,
then some artifact may appear twice on different pagination windows.

In order to mitigate the risk, we should check if an artifact has
already be exported before exporting it again.

Change-Id: I4109e9e17cf782587af44d44330a8c9c2302f0fb