Simplify types definitions for comments presenter
part of story #31091 Add the hybrid comment to manage comments

This contribution aims to simplify the types definitions for the
different kinds of PullRequestPresenters.

How did we get here?

We have several types of items to display as comments having different shapes:
- timeline items: GlobalComment vs CommentOnFile vs ActionOnPullRequestEvent
- file-diff inline comments

We do not want to display them the same everywhere:
- In the timeline we display the file in the header for CommentOnFiles
- In the file-diff view, we don't
- An ActionOnPullRequestEvent cannot be a thread

Some info are required to be able to save a reply:
- unidiff_offset, file_path, position etc.
- user_id

How to reduce the complexity?

Since we can differentiate the different types of comments by comparing
their types with the list of SupportedTimelineItemTypes, we can get rid
of is_inline_comment.

By reworking the type CommonComment, we are now able to have cleaner
presenters:

- file is now an exclusive attribute of the
  PullRequestInlineCommentPresenter type, it is not optional anymore

- We no longer have unidiff_offset, position, file_path and is_file_diff_comment
  polluting the types. Before this patch, there were too much
  "file-diff" logic in the component. Since the comment component is in
  a shared lib, there is no reason to keep it in there. Now, to display
  a comment in the file-diff view, we build a PullRequestInlineCOmmentPresenter, and
  that's it.

In the meantime, I've added some additional steps after a reply has been
saved: When we receive the payload from the REST route, it does not have
necessarily the right form, and a conversion from a NewComment to a
PullRequestComment is needed to avoid having presenters with empty
attributes.

How to test:
- No functional changes expected in:
--> The angular overview
--> The vue overview
--> The file-diff view in sbs AND unidiff modes

Which means:
- In overviews:
--> comments on files have the file displayed
--> The link to the comment in the file works as usual
--> You can reply to any comment
--> You can post global comments
--> Pullrequests events are displayed properly

- In the [Changes] tab
--> You can post new comments
--> The comments have the right place
--> The diffs are displayed properly

Change-Id: I2de072056e19fd3667fe5f512d5466a461ee3482