Fix request #31136 Jira board by ID does not retrieve board project
In some cases, the Jira REST API for board does not retrieve the location of a board.
We will have to get the project using another API call to be able do to the verification

According to documentation[0], location key might not be here.
We have to use agile/1.0/board/{boardId}/project instead.

[0] https://docs.atlassian.com/jira-software/REST/7.3.1/#agile/1.0/board-getBoard

Change-Id: Ic604f2c196ecae7ee297a4e5773dac65d7438527