feat: temporary page for MW under construction
This introduces a dedicated page to display an empty state for MediaWiki
standalone service.

This will be used later when a project is being migrated to the new
service, because migration is taking some time and we should display
this temporary page until the migration is complete.

To access the page:
https://tuleap.example.com/mediawiki_standalone/under-construction/<project-unix-name>

Part of story #26041: migrate mediwiki from 1.23 to 1.35.X

Change-Id: I97b6d75132b54b459001a0ff8f6722682501174a