Display pull request description as "first comment"
part of story #31143 Add the pull-request description as first possible comment

This contribution introduces a new comment component to manage the
display of the pull-request description. Because it is displayed as a
comment, while it is not at all a comment, I've decided to have a
dedicated component.

How to test:
- Go to a pull request having a description
--> The description is displayed as a comment, at the very start of the
    timeline.
--> You can't edit it (yet) nor add a reply to it (never).

- Go to a pull-request without any description where you have the WRITE
  permission
--> The first comment displays an empty state text saying that there is
    no description yet please provide one kindly asap.

- As another user without the WRITE permission, go to the same pull-request
--> Same expectation than the previous one, expect that you don't have
    the "Please add one" part.

Change-Id: I46591c05f78b4ad287fcd0c3ca917e7d6339277a