feat: Edit link type for new links
How to test:
- No functional change for the link type selectbox that lets you add new
links (it still works)
- after adding a new link, its type is now an editable selectbox
- when you change the new link's type, it is saved correctly in backend
- you cannot choose `_mirrored_milestone` type
- when the artifact already has a parent (when you edit an existing task
OR when you add a new task to a story in Planning), then "is Child of"
type will be disabled
- after adding a new parent ("is Child of" type), if change its type
(for example to "is Linked to"), then you can once again add a new
parent.
You should not be able to add two parents by changing the type of links.

Notes:
Having a full element for the types selector means it is easier to reuse
it. It keeps the same logic but depending on where it is used, the
onChange and the value will be different.

The SCSS classes will be merged again in a later contribution when we
can edit existing link types too.

part of story #24970 [modal] select type for links

Change-Id: Iddb2edbc44b260157f1c51bc6e6c7a0ca954937d