fix: throws error when a reverse link cannot be removed
Part of story #24970 [modal] select type for links

Throws an error instead of ignoring the error when we try to remove a
reverse link which cannot be removed. Currently, only '_is_mirrored_'
type is concerned

To test you can try to remove a `is_mirrored` link type of an artifact from a Team project.

Payload example:

```
{
  "values": [
    {
      "field_id": 4428,
      "all_links": [
        {
          "id": 160,
          "direction": "reverse",
          "type": "_is_child"
        }
      ]
    }
  ]
}
```

Change-Id: Ibf41a32edf2e0f9f72c128c81d0d6ab60b65785f