fix: Unable to open the edit title/description modal anymore
part of part of story #31151 Edit the description comment in text only

Looks like it is not possible to inject $element in a TlpModal
controller, nor the TlpModal service. This fix makes the
overview-controller give the modal a on_success_callback to trigger when
the title and description have been successfully saved.

This callback loads tooltips present in the area where the description
is displayed, to avoid having several instances of the tooltips for
every single reference.

Change-Id: I33059db0c3897090b4ee8ab2207f792762b9cf31