refactor: split autocompleter out of controller
No functional change. Searching for artifacts in the link field still
functions as before.

Notes:
Splitting the two is needed, so that we can DRY some of the shared code.
For example, only the controller should know how to retrieve Possible
parents. It should not be duplicated.

part of story #24970 [modal] select type for links

Change-Id: I9c11cc0136463153693e9ce88ee8d3ea33d8f638