refactor: autocompleter should use controller to retrieve parents
No functional change in the Possible parents autocompletion. When you
choose the "is Child of" type of link, it still shows a list of possible
parents. You can still filter that list by typing in the link selector
field.

Notes:
The retrieval of parents was previously duplicated, because the
controller needed to know if there were possible parents before
pre-selecting the type "is Child of". It is now de-duplicated. Also, the
autucompleter alone builds a GroupOfItems, the controller does not know
about it, it returns Domain objects.

part of story #24970 [modal] select type for links

Change-Id: I59794d34ef85078a6a54dc86caf604e5485b4f8f