Have a common writing zone + manage focus state
part of part of story #31151 Edit the description comment in text only

The goal of this contribution is to have a common writing zone between
the components. In the mockups, the writing zone has tabs above the
textarea, showing [Write] and [Preview]. Since the implementation of
markdown will be done in another phase of the epic, only the [Write] tab
is shown atm.

This writing zone, when it is "active" should have its
textarea focused, the whole content should have its border in
tlp-main-color, and the tab [Write] should be active. In the contrary,
when it is not active, the borders should be in neutral color, the tab
inactive, and the textarea not focused.

As per the mockups, this writing zone is expected to have the same
design and behavior in the three components: the new-comment for, the
reply form and the description comment edition form.

During the development, I realized that the fact NewCommentForm had no
controller made the usage of the writing zone too messy, so I've decided
to split the component in different files, like it's done in the other
components. It allows us to have better unit tested parts, and a cleaner
public api for the component.

Regarding the post_submit_callback, I had to expose a method in the
controller to trigger it manually, only for testing purpose. I know it
doesn't feel quite right, but it was the easiest and cleanest solution to avoid breaking the tests in SideBySideCodeMirrorWidgetCreator.

How to test:
- In the vue and angular overview:

-- Click [Edit] on the description comment
---> The writing zone is active
---> It is inactive when you click outside

-- Click [Reply] on a comment
---> The writing zone is active
---> It is inactive when you click outside

- In the vue overview only
-- Scroll down to the "add comment component"
--> It is inactive by default
--> Clicking in the textarea or the [Write] tab make it active
--> Clicking outside makes it inactive
--> When you save a new comment, it is added to the timeline, the
textarea is cleared and the writing zone is in inactive mode.

- In the file-diff view
-- Click to add a comment
--> The writing zone is active and the textarea focused by default

Change-Id: Iec2a6b1da354536b3b608960d3ae60d14536ac29