Enable back controls when comment couldn't be saved
part of part of story #31151 Edit the description comment in text only

When an error occured while saving a comment or the description comment,
the controls are still disabled, preventing the user to have a chance to
fix the error and to submit again.

Now when an error occures, we reset is_being_submitted back to false in
the form presenters, making the buttons enabled again.

How to test:
- Throw RestExceptions in PullRequestsResource l.731, l.913 and l.1143
- Edit the description comment and submit
--> The error modal opens
--> The [Cancel] and the [Save] buttons are not disabled anymore

- Create a reply to an existing comment
--> The error modal opens
--> The [Cancel] and the [Reply] buttons are not disabled anymore

- Write a global comment and submit it
--> The error modal opens
--> The [Comment] button is not disabled anymore

Change-Id: I6527097254ca8d65788141738a41709955b072c7