feat: link-selector opens on pointerup
How to test:
No major functional change in link-selector. It opens when releasing the
pointer instead of when pressing it.

Notes:
The previous behaviour meant that it was possible to accidentally select
an item just at the same time as you opened up the dropdown, when
keeping the pointer pressed down. It appeared in cypress tests while
clicking the selector element to open the dropdown.

part of story #24970 [modal] select type for links

Change-Id: I17649f5a71832e5f292e385a95e719d80a3441c8