fix: writing zone header messes components corners
part of part of story #31151 Edit the description comment in text only

When the writing zone is displayed, its top right corner messes with
it's parent's border radius, making "a hole" in it. It is really small,
but yet enough to be spottd by our beloved UXD :')

Sadly, I could not use an `overflow: hidden` rule on the
pull-request-comment class, otherwise the arrow displayed on it's left
is hidden too.

How to test:
- In the description comment edition form
- In the reply form
- In the new-comment form

--> No matter if the writing zone is active or not, the border of the
    component is nice and smooth/continous

Change-Id: I1e19426468b394dd30c040f3d301e915d3abbb91