feat: Display list of project to migrate
As siteadmin I have a new siteadmin entry "MediaWiki standalone"
So that I can see the list of active projects that are using legacy
mediawiki, and the status of the ongoing migration, if any.

If a project has already started a migration, a badge info is displayed
and the button to start the migration is disabled.

If a project migration ended in error, a badge danger is displayed and
the button to start the migration is disabled.

If a project is not allowed to use the plugin, then a badge is displayed
and the button to start the migration is disabled.

If a project has activated the legacy service but never initialized it,
then it is not listed at all for now, it deserves a special treatment
in a dedicated contribution.

The table can be filtered with the project name or the status of the
migration.

For now the button to start the migration does nothing, scheduling of
the migration will be part of a dedicated contribution.

Part of story #26041: migrate mediwiki from 1.23 to 1.35.X

Change-Id: I21d131be70c03c7c7225a6e4d57f8c3b89e526e0