Drop incomplete MFA (TOTP based) plugin
This plugin has never been finished and it is unlikely it will ever be in this state.
There is some incoming WebAuthN work which is better than this anyway.

Worst case scenario we could still get back the code thanks to the Git history
but it is not really up to our current standards.

Close story #11649: protect my account with a TOTP code

Change-Id: I0a45e71a0c264bcbc7c372fffcf329774068ea28