Deal with pending 'single DB migration' before migrating to MW Standalone
MW Standalone supports both mode but it is either one of the other, it
cannot mix the two.
When using the 'single DB' mode, we need to to make sure MW instances
are using this mode before trying to move them to MediaWiki standalone.

Part of story #26041: migrate mediwiki from 1.23 to 1.35.X

Change-Id: I53a883f70ec971b99233a98f826f8a0f3c6dffb4