fix: Create labels on the fly when doing Jira Import
Jira Server doesn't have an API to list labels. However it turns out
that there is no need for having the whole list of labels upfront
because we can rely on OpenList, on the fly creation of values.

As downside, it's not possible to import labels that are defined but not
used but it's an acceptable tradeoff (it might even not be possible in
Jira).

Closes request #31120 Jira Labels field must be imported as an open list

Change-Id: Id03932fb07801a61fdbac17ec860e8dcb5238b58