fix: title and description should be patchable individually
part of story #31151 Edit the description comment in text only

When we want to update the pull-request title or description, we need to
update both at the same time. We should be able to update them
individually. Currently, there is a bug in pullreqsuest-overview where
the title or the description is reset if you update the title or the
description in the meantime.

With this fix, it is now possible to update them individually.

How to test:
- In the angular app, open the update modal and update both the title
  and the description
- submit

--> They both have been updated within the same PATCH request

- In the vue overview, update the title
- Update the description
- Refresh the page

--> The title is still the new one

- Update the title once again
- Refresh the page

--> The description hasn't been reset

Change-Id: I5f244fadae83e9a1b65c0b1968884f28b5c629c5