feat: Redirect to new MW
If user tries to access to a legacy MediaWiki page after it has been
migrated by the siteadmin, then it currently gets an error page
"Mediawiki service is not active in this project".

Instead I suggest to redirect to the new service instead.

Part of story #26041: migrate mediwiki from 1.23 to 1.35.X

Change-Id: Id17c561a32b5d9fdf9b459eef34c893af97d4aeb