Package the Wasmtime lib wrapper
To test, build the package with `nix-build ./src/additional-packages/tuleap-wasmtime-wrapper-lib.nix`
then install on your dev instance.
You should find the library and the header in `/usr/lib/tuleap/wasm/`

The build is done without `cargo-zigbuild` as it was causing build issues in some environment.
It is possible to build the library in the development nix-shell with
`cargo-build-glibc217 --release`

Part of story #31077: deploy a WebAssembly module to validate incoming references as a pre-receive hook

Change-Id: I481f99702c9e590c78d1675f84b7fa1012513409