Make possible to have "additional packages" specific to an OS
It is not needed to have the mailman-tuleap and cvs-tuleap packages in
the EL9 repository or the have the ViewVC package for Python 3 in the
EL7 repo.

This limit the chances of ending up in situation where packages are
installed where they should not and conflicts between packages.

Part of: request #29619 Add support of Enterprise Linux 9

Change-Id: Ie3b66ba20a208781d138076c8d631790c91cc8cb