feat: Display an "Already merged" state
part of story #31186 Add management of "Merge"

This contribution introduces an "already merged" state which aims to be
displayed when the pull-request is already merged.

The pull-request representation now contain a "status_info" key which
contain the date, status and user who has merged it.

For the moment, I do not take into account the "abandoned" state, which
will be implemented in a dedicated user story.

How to test:
- Go to an open pull-request
- Merge it
--> The state is displayed with your avatar and the merge date.

Change-Id: I6909fb1d0e34404780d8a5fc32b072c66995824b