design: Don't display a border below PR infos when user cannot write
part of story #31186 Add management of "Merge"

As per the mockups, when the PR is open and the user has not the right
to merge the pull-request, the component containing this button is not
displayed. However, there should not be a border below the PR infos if
the component is hidden.

How to test:
- Go to a pull-request with a user only having READ permissions in the
  repository

--> No border is displayed below the PR info section

Change-Id: Ia6b478be97006c12ced7aad509eea35fd69ecb84